Remove notion client class
